requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
alembic==1.11.1
psycopg[binary]==3.3.3
python-dotenv==1.0.0
Werkzeug==3.1.3
requests==2.32.3